+++ KVLite.UnitTests/NinjectConfig.cs
@@ -25,9 +25,9 @@ using CodeProject.ObjectPool;
 using CodeProject.ObjectPool.Specialized;
 using Common.Logging;
 using Common.Logging.Simple;
-using Finsa.CodeServices.Clock;
-using Finsa.CodeServices.Compression;
-using Finsa.CodeServices.Serialization;
+using PommaLabs.CodeServices.Clock;
+using PommaLabs.CodeServices.Compression;
+using PommaLabs.CodeServices.Serialization;
 using Ninject.Modules;
 
 namespace PommaLabs.KVLite.UnitTests