+++ KVLite.UnitTests/VolatileCacheTests.cs
@@ -21,8 +21,8 @@
 // DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using Finsa.CodeServices.Caching;
-using Finsa.CodeServices.Clock;
+using PommaLabs.CodeServices.Caching;
+using PommaLabs.CodeServices.Clock;
 using Ninject;
 using NUnit.Framework;
 using PommaLabs.KVLite.Core;